+++ _tests_/react-ts-tw/src/constants/index.ts
@@ -1,3 +1,5 @@
 // Test: index-export-style (shorthand exports in index files)
 
+export { commonConstants } from "./common";
 export { localeConstants } from "./locale";
+export { routesConstants } from "./routes";